% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hextile.R
\name{attach_hex_dependency}
\alias{attach_hex_dependency}
\title{Attach hex.css dependency to hex_wall}
\usage{
attach_hex_dependency(tag)
}
\arguments{
\item{tag}{An htmltools tag object}
}
\value{
The tag with hex.css dependency attached
}
\description{
Attach hex.css dependency to hex_wall
}
\keyword{internal}
