% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hextile.R
\name{hex_dependency}
\alias{hex_dependency}
\title{Include hex.css dependency}
\usage{
hex_dependency()
}
\value{
An htmltools HTML dependency object
}
\description{
Includes the hex.css file as an HTML dependency
}
