% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hextile.R
\name{hextile}
\alias{hextile}
\title{Create a hexagon tile pattern}
\usage{
hextile(images, class = NULL, ...)
}
\arguments{
\item{images}{Character vector of image paths/URLs}

\item{class}{Additional CSS classes to apply to the container}

\item{...}{Additional attributes for the container div}
}
\value{
An htmltools tag object
}
\description{
Creates a hexagonal tile pattern using hex.css styling. Each image is displayed
in a hexagonal shape arranged in a tiling pattern.
}
\examples{
\dontrun{
# Simple usage with image paths
hextile(c("logo1.svg", "logo2.svg", "logo3.svg"))

# With additional CSS classes
hextile(c("img1.png", "img2.png"), class = "my-custom-class")
}
}
